Push Zuora REST models into zuora service package
